% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmn-loglik.R
\name{lmn.loglik}
\alias{lmn.loglik}
\title{Log-likelihood function for LMN models.}
\usage{
lmn.loglik(Beta, Sigma, suff, Y, X, V, acf)
}
\arguments{
\item{Beta}{an (p x q) matrix (see details).}

\item{Sigma}{an (q x q) matrix (see details).}

\item{suff}{result of call to lmn.suff (optional, to avoid calculations).}

\item{Y}{an (n x q) matrix.}

\item{X}{an (n x p) matrix. May also be passed as
\itemize{
   \item \code{X = 0}: in which case there is no intercept,
   \item \code{X != 0}: in which case a scaled intercept X = X * matrix(1, n, 1)
   is assumed.
}}

\item{V}{either: (1) an \code{n x n} full matrix, (2) an vector of length \code{n} such that \code{V = diag(V)}, (3) a scalar, such that \code{V = V * diag(n)}.}

\item{acf}{a vector of length n such that V = toeplitz(acf).}
}
\value{
The calculated log-likelihood.
}
\description{
Log-likelihood function for LMN models.
}
\details{
The model is defined as
\deqn{Y ~ MNorm(X*Beta, V, Sigma),}
where MNorm is the Matrix-Normal distribution, i.e.
\deqn{vec(Y) ~ N( vec(X*Beta), Sigma \otimes V ).}
}
\examples{
## Data
Y = matrix(rnorm(100),50,2)

## Scalar V example
X = matrix(1,50,1)
q = ncol(X)
V = 0.5
suff = lmn.suff(Y, X, V=V)
Beta = matrix(c(1,0.5),1,2)
Sigma = diag(2)
lmn.loglik(Beta, Sigma, suff, Y, X, V)

}
