% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmn.R
\docType{package}
\name{LMN}
\alias{LMN}
\alias{LMN-package}
\title{(L)inear (M)odel for (N)uisance Parameters}
\description{
Efficient profile likelihood and marginal posteriors when nuisance parameters are those of linear regression models.
}
\details{
The methodology applied to the parameters \code{theta, Beta, Sigma} of models of the form
\deqn{Y | theta ~ MN( X(theta) * Beta, V(theta), Sigma).}
A conditionally conjugate prior for (\code{Beta}, \code{Sigma}) is
\deqn{(Beta, Sigma) | theta ~ MNIW(Lambda, Omega, Psi, nu).}
Note that the hyperparameters can depend on \code{theta}, e.g., \code{Lambda(theta)}, \code{Omega(theta)}, etc.  To achieve some of the restricted forms of the model:
\itemize{
\item \code{Omega = 0}: Flat prior \code{Beta ~ 1}.
\item \code{Psi = 0}: Improper prior \code{Sigma ~ 1/sqrt(|Sigma|)^(v+p+1)}.
\item \code{Omega = NA}: No \code{Beta}, i.e., Y ~ MN(0, V, Sigma).
\item \code{nu = NA}: No \code{Sigma}, i.e., Y ~ MN(X * Beta, V, I).
}
}
