% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmn-suff.R
\name{lmn.suff}
\alias{lmn.suff}
\title{Calculate the sufficient statistics of an LMN model.}
\usage{
lmn.suff(Y, X, V, Vtype, npred = 0)
}
\arguments{
\item{Y}{An \code{n x q} matrix of responses.}

\item{X}{An \code{N x p} matrix of covariates, where \code{N = n + npred} (see \strong{Details}). May also be passed as:
\itemize{
  \item A scalar, in which case the one-column covariate matrix is \code{X = X * matrix(1, N, 1)}.
  \item \code{X = 0}, in which case the mean of \code{Y} is known to be zero, i.e., no regression coefficients are estimated.
}}

\item{V, Vtype}{The between-observation variance specification.  Currently the following options are supported:
\enumerate{
  \item \code{Vtype = "full"}: \code{V} is an \code{N x N} symmetric positive-definite matrix.
  \item \code{Vtype = "diag"}: \code{V} is a vector of length \code{N} such that \code{V = diag(V)}.
  \item \code{Vtype = "scalar"}: \code{V} is a scalar such that \code{V = V * diag(N)}.
  \item \code{Vtype = "acf"}: \code{V} is either a vector of length \code{N}, or an object of class \code{Toeplitz} such that \code{V = toeplitz(V)}.
}
For \code{V} specified as a matrix or scalar, \code{Vtype} is deduced automatically and need not be specified.}

\item{npred}{A nonnegative integer.  If positive, calculates sufficient statistics to make predictions for new responses. See \strong{Details}.}
}
\value{
An S3 object of type \code{lmn_suff}, consisting of a list with elements:
\itemize{
  \item \code{Beta.hat = (t(X)V^{-1}X)^{-1}t(X)V^{-1}Y}, or \code{NULL} if \code{X = 0}.
  \item \code{T = t(X)V^{-1}X}.
  \item \code{S = t(Y-X*Betahat)V^{-1}(Y-X*Betahat)}.
  \item \code{ldV = log(|V|)}.
  \item \code{n = nrow(Y)}.
  \item \code{p = nrow(Beta)}, or \code{p = 0} if \code{X = 0}.
  \item \code{q = ncol(Y)}.
}
In addition, when \code{npred > 0}:
\itemize{
  \item \code{Ap = t(w)V^{-1}Y}.
  \item \code{Xp = x - t(w)V^{-1}X}.
  \item \code{Vp = v - t(w)V^{-1}w}.
}
}
\description{
Calculate the sufficient statistics of an LMN model.
}
\details{
The multi-response normal linear regression model is defined as
\preformatted{
Y ~ MNorm(X Beta, V, Sigma),
}
where MNorm is the Matrix-Normal distribution, i.e., corresponding to the multivariate normal distribution
\preformatted{
vec(Y) ~ N( vec(X Beta), Sigma %x% V ),
}
where \code{vec(Y)} is a vector of length \code{n*q} stacking the elements of \code{Y}, and \code{\%x\%} is the Kronecker product.  The function \code{lmn.suff} returns everything required to calculate the (normalized) probability distribution \code{p(Y | Beta, Sigma, X, V)}.  For mathematical details please see \code{vignette("LMN", package = "LMN")}.

When \code{npred > 0}, summary statistics are produced to calculate the conditional (predictive) distribution \code{p(y | Y, XX, VV, Beta, Sigma)}, where
\preformatted{
YY = rbind(Y, y) ~ MNorm(XX*Beta, VV, Sigma),
}
\code{VV = rbind(cbind(V, w), cbind(t(w), v))} and \code{XX = rbind(X, x)}.  In this case the inputs to \code{lmn.suff} are \code{Y}, \code{V = VV} and \code{X = XX}. Please see vignette for mathematical details.
}
\examples{
# Data
n <- 50
q <- 3
Y <- matrix(rnorm(n*q),n,q)

# No intercept, diagonal V input
X <- 0
V <- exp(-(1:n)/n)
lmn.suff(Y, X = X, V = V, Vtype = "diag")

# X = (scaled) Intercept, scalar V input (no need to specify Vtype)
X <- 2
V <- .5
lmn.suff(Y, X = X, V = V)

# X = dense matrix, Toeplitz variance matrix
p <- 2
X <- matrix(rnorm(n*p), n, p)
acf <- SuperGauss::Toeplitz(acf = 0.5*exp(-seq(1:n)/n))
lmn.suff(Y, X = X, V = acf, Vtype = "acf")

}
