% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmn-prof.R
\name{lmn.prof}
\alias{lmn.prof}
\title{Profile Log-Likelihood for the LMN Model}
\usage{
lmn.prof(suff, Y, X, V, acf, noSigma = FALSE)
}
\arguments{
\item{suff}{result of call to lmn.suff (i.e. avoid calculations here)}

\item{noSigma}{logical. if true assumes that \code{Sigma = diag(ncol(Y))}.}
}
\value{
the calculated profile log-likelihood.
}
\description{
Profile Log-Likelihood for the LMN Model
}
\details{
The model is defined as
\deqn{Y ~ MNorm(X*Beta, V, Sigma),}
where MNorm is the Matrix-Normal distribution, i.e.
\deqn{vec(Y) ~ N( vec(X*Beta), Sigma \otimes V ).}
}
